requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.0
markdown==3.5.1
chromadb==0.4.18
pydantic==2.5.0
uuid==1.30
python-multipart==0.0.6
streamlit-chat==0.1.1
plotly==5.17.0